aws-test: Drop need for parameters, just use latest/
Get rid of the need to pass parameters to the `aws-test` job in
preparations for turning on `keepUndefinedParameters=true` in the
Jenkins instance.

We only every run the `aws-test` job from the `cloud` job, so we know at
that point that the latest output is in `latest/`. We can just resolve
that back to a specific versioned dir before proceeding.

This normalizes the `aws-test` job to take the exact same parameters as
the others, and makes it easy to e.g. run the job manually like the
others.