extensions/Dockerfile: only call get-ocp-repo.sh in OpenShift CI
We want to be able to build this container directly via `podman build`
in production, but the way this conditional is structured breaks things
because we're always calling `get-ocp-repo.sh`, which is only useful
when building in OpenShift CI.

Really, the conditional there was not correct. We should check for
`OPENSHIFT_CI` instead like in the main `Containerfile` for the
node image, and set that same build arg in openshift/release. The
`--create-gpg-keys` case was needed when building `okd-c9s` in CI, but
we don't do that anymore, so drop it and drop the switch from the script
since it no longer has a user.