extensions/Dockerfile: add extensions.json to last layer
Just like 6241505 ("Containerfile: add metadata in last layer of node
image"), we want to make it easy for anyone but particularly ART to
query metadata about the extensions container.

Instead of generating `extensions.json` but assuming it's a bind-mount,
directly put it in the final container image. In the cosa path, we'll keep
propagating that info up to the build's `meta.json` to not regress, but
now at least in the container-native flow, we also have that metadata
accessible either by doing an `oc image extract` or a
`podman run $img cat /path/to/extensions.json`.