extensions: use baked-in treefile
The extensions container build already builds from the node image, which
itself builds from the base CoreOS image, which means that the manifest
file of that original base compose is still there. So let's reference
that location instead.

That allows us to sever the extensions build from files that define
the base compose. This is prep for splitting the latter into a separate
repo.